Moved get_db to dependencies file
