+++ src/application/backend/app/routers/auth.py
@@ -3,7 +3,7 @@ from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 from datetime import timedelta
 
-from app.db.session import get_db
+from app.dependencies import get_db
 from app.services.auth_service import authenticate_user, create_access_token
 from app.core.security import ACCESS_TOKEN_EXPIRE_MINUTES
